+++ R/confint_stdmod.R
@@ -6,7 +6,7 @@
 #' @details If bootstrapping is used to form the confidence interval by
 #' [std_selected_boot()],
 #' users can request the percentile confidence intervals of
-#' the bootstrap estimates.
+#' the bootstrap estimates. This method does not do the bootstrapping itself.
 #'
 #' @return
 #'  A matrix of the confidence intervals.
@@ -17,13 +17,14 @@
 #'            If missing, the confidence intervals of all parameters will be
 #'            returned.
 #' @param level The level of confidence. For the confidence intervals returned
-#'             by [lm()], default is .95. For the bootstrapping
+#'             by [lm()], default is .95, i.e., 95%. For the bootstrap
 #'             percentile confidence
 #'             intervals, default is the level used in calling
 #'             [std_selected_boot()]. If a level different from that in the
 #'             original
 #'             call is specified, `full_output` needs to be set in the call
-#'             to [std_selected_boot()] such that the original bootstrap output
+#'             to [std_selected_boot()] such that the original bootstrapping
+#'             output
 #'             is stored.
 #' @param type The type of the confidence intervals. Default is `"lm"`,
 #'            returned by the [confint()] method of [lm()]. If set to `"boot"`,
@@ -37,8 +38,7 @@
 #' @examples
 #'
 #' # Load a sample data set
-#' # It has one predictor (iv), one moderator (mod), on covariate (v1),
-#' # one categorical covariate (cat1) with three groups, and one dv (dv).
+#' 
 #' dat <- test_x_1_w_1_v_1_cat1_n_500
 #'
 #' # Do a moderated regression by lm
@@ -50,6 +50,7 @@
 #' lm_std <- std_selected(lm_raw, to_scale = ~ .,
 #'                                to_center = ~ .)
 #' summary(lm_std)
+#'
 #' confint(lm_std)
 #'
 #' # With bootstrapping
@@ -59,8 +60,11 @@
 #'                                          to_center = ~ .,
 #'                                          nboot = 100)
 #' summary(lm_std_boot)
+#'
 #' confint(lm_std_boot)
+#'
 #' # Bootstrap percentile intervals
+#'
 #' confint(lm_std_boot, type = "boot")
 #'
 #' @export