+++ R/stdmod.R
@@ -6,36 +6,42 @@
 #' @details Two more general functions, [std_selected()] and
 #' [std_selected_boot()], have been developed and can do what these functions
 #' do and more. Users are recommended to use them instead of [stdmod()] and
-#' [stdmod_boot()].
+#' [stdmod_boot()]. These two functions will not be updated in the near
+#' future.
 #'
 #' Nevertheless, if computing the standardized moderation effect and
 #' forming its nonparametric
-#' bootstrap interval are what users need, then these functions can still
+#' bootstrap interval are all required, then these functions can still
 #' be used.
 #'
 #' [stdmod()] computes the standardized moderation effect given an
 #' [lm()] output using the formula from Cheung, Cheung, Lau, Hui, and Vong
 #' (2022). Users specify
-#' the moderator, the predictor, the outcome variable, and the corresponding
+#' the moderator, the focal variable (the variable with its effect on
+#' the outcome variable moderated), the outcome variable (dependent variable)
+#' , and the corresponding
 #' standardized moderation
 #' effect. Users can also select which variable(s) will be standardized.
 #'
 #' @return
-#' The standardized moderation effect.
+#' [stdmod()] returns a scalar: The standardized moderation effect.
 #'
 #' @param lm_out The output from [lm()].
-#' @param x      The independent variable, that is, the variable with its effect
+#' @param x      The focal variable, that is, the variable with its effect
 #'              being moderated. If supplied, its standard deviation will be
 #'              used
-#'              for rescaling. Default is `NULL`.
+#'              for rescaling. Also called the independent variable in some
+#'              models. Default is `NULL`.
 #' @param w      The moderator. If supplied, its standard deviation will be
 #'              used
 #'              for rescaling. Default is `NULL`.
-#' @param y      The dependent (outcome) variable. If supplied, its standard
+#' @param y      The outcome variable (dependent variable) . If supplied,
+#'              its standard
 #'              deviation will be used for rescaling. Default is NULL.
 #' @param x_rescale  If `TRUE`, will rescale x by its standard deviation.
 #'                   Default is `TRUE`.
-#' @param w_rescale  If `TRUE`, will rescale w by its SD. Default is TRUE.
+#' @param w_rescale  If `TRUE`, will rescale w by its standard deviation.
+#'                    Default is `TRUE`.
 #' @param y_rescale  If `TRUE`, will rescale y by its standard deviation.
 #'                   Default is `TRUE`.
 #'
@@ -51,8 +57,7 @@
 #' @examples
 #'
 #' # Load a test data of 500 cases
-#' # It has one predictor (iv), one moderator (mod), two covariates (v1 and v2),
-#' # and one dv (dv). All variables continuous.
+#'
 #' dat <- test_x_1_w_1_v_2_n_500
 #'
 #' # Do regression as usual: