+++ R/vcov_stdmod.R
@@ -11,7 +11,7 @@
 #' @author Shu Fai Cheung <https://orcid.org/0000-0002-9871-9448>
 #'
 #'@return
-#'  A matrix of the variances and covariances.
+#'  A matrix of the variances and covariances of the parameter estimates.
 #'
 #'@param object The output of [std_selected()] or [std_selected_boot()].
 #'@param type The type of variance-covariance matrix. Default is `"lm"`,
@@ -23,20 +23,17 @@
 #'
 #'@examples
 #' # Load a sample data set
-#' # It has one predictor (iv), one moderator (mod), on covariate (v1),
-#' # one categorical covariate (cat1) with three groups, and one dv (dv).
+#'
 #' dat <- test_x_1_w_1_v_1_cat1_n_500
 #' head(dat)
 #'
 #' # Do a moderated regression by lm
 #' lm_raw <- lm(dv ~ iv*mod + v1 + cat1, dat)
-#' summary(lm_raw)
 #'
 #' # Standardize all variables except for categorical variables.
 #' # Interaction terms are formed after standardization.
 #' lm_std <- std_selected(lm_raw, to_scale = ~ .,
 #'                                to_center = ~ .)
-#' summary(lm_std)
 #'
 #' # VCOV of lm output
 #' vcov(lm_std)
@@ -49,7 +46,6 @@
 #'                                  conf = .95,
 #'                                  nboot = 100)
 #' # In real analysis, nboot should be at least 2000.
-#' summary(lm_std_boot)
 #'
 #' # VCOV of lm output
 #' vcov(lm_std_boot)